+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.4 (WIP)
 
 * Use `he-context-menu` element to show widgets to add
+* Add widget actions to move across container boundaries
+
 ## 0.1.3 (2022-03-21)
 
 * Catch and print errors when appending a widget to a container